Test code for subsampling for easier testing with less data. To be removed. It is disabled as is so has no effect. To enable edit the define on line 250 in source_driver_ros1.hpp

+++ src/manager/source_drive_common.hpp
@@ -80,7 +80,7 @@ public:
                            driver_param.decoder_param.remake_config.use_ring_for_vertical, false);
             YamlRead<bool>(driver_config["remake_config"], "duplicate_sparse_rings",
                            driver_param.decoder_param.remake_config.duplicate_sparse_rings, false);
-            YamlRead<int>(driver_config["remake_config"], "echo_mode_filter",
+            YamlRead<uint16_t>(driver_config["remake_config"], "echo_mode_filter",
                           driver_param.decoder_param.echo_mode_filter, 0);        }
         // ROS related
         YamlRead<bool>(       config["ros"], "send_packet_ros",            driver_param.input_param.send_packet_ros, false);
